requirements.txt
flask==2.2.2
flask_sqlalchemy==2.5.1
python-dotenv==0.20.0
psycopg2=2.9.3